% Generated by roxygen2 (4.0.1): do not edit by hand
\name{alignment_depth}
\alias{alignment_depth}
\title{Plot alignment depth.}
\usage{
alignment_depth(bamfile, reffile)
}
\arguments{
\item{bamfile}{Path to the bam file.}

\item{reffile}{Path to the reference sequence (Fasta format).}
}
\value{
Side effects. Graphs alignment depth as a line-plot and summarises the alignment depth in a
boxplot.
}
\description{
Plot alignment depth.
}

