% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{cigar_string}
\alias{cigar_string}
\title{Return the CIGAR string from a \code{bamAlign} object.}
\usage{
cigar_string(align)
}
\arguments{
\item{align}{A \code{\linkS4class{bamAlign}} object}
}
\value{
A CIGAR string.
}
\description{
Return the CIGAR string from a \code{bamAlign} object.
}

