% Generated by roxygen2 (4.0.1): do not edit by hand
\name{consensus_table}
\alias{consensus_table}
\title{Generate a consensus table between a BAM alignment and a reference sequence.}
\usage{
consensus_table(bamfile, reffile)
}
\arguments{
\item{bamfile}{Path to the BAM file.}

\item{reffile}{Path to the reference sequence (Fasta format).}
}
\value{
A data.table. See \code{Details}.
}
\description{
Generate a consensus table between a BAM alignment and a reference sequence.
}
\details{
\code{consensus_table} generates a \code{\link{data.table}} with 5 columns:
The \emph{Position} along the reference, the \emph{Nucleotide} (A, C, G, T, or Gap [-]),
the number of reads at a Position for each Nucleotide, the
}

