% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{porequal}
\alias{porequal}
\alias{porequal-package}
\title{Assessing the quality of high-throughput multiple sequence alignments.}
\description{
This package provides functions to perform graphical quality assessment of
SAM/BAM file-based multiple alignments.
}
\details{
\tabular{ll}{
   Package: \tab porequal\cr
   Type: \tab Package\cr
   Version: \tab 0.0.1\cr
   Date: \tab 2014-09-04\cr
   License: \tab MIT\cr
 }
}
\author{
Gerhard Schöfl \email{gerhard.schofl@gmail.com}

Maintainer: Gerhard Schoefl \email{gerhard.schofl@gmail.com}
}
\keyword{package}

