% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_fasta}
\alias{read_fasta}
\title{Read a fasta file.}
\usage{
read_fasta(path, as = c("character", "DNAStringSet"))
}
\arguments{
\item{path}{Path to fasta file.}

\item{as}{Return sequence as a character vector or as a
\code{\linkS4class{DNAStringSet}} object.}
}
\value{
A named character vector or a \code{\linkS4class{DNAStringSet}} object.
}
\description{
Read a fasta file.
}

