% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sam2bam}
\alias{sam2bam}
\title{Convert a SAM file to a BAM file.}
\usage{
sam2bam(samfile, reffile)
}
\arguments{
\item{samfile}{Path to the SAM file.}

\item{reffile}{Path to the reference sequence (Fasta format)}
}
\value{
The paths to the BAM file.
}
\description{
Convert a SAM file to a BAM file.
}

