% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{consensus_table}
\alias{consensus_table}
\title{Generate a consensus table between a BAM alignment and a reference sequence.}
\usage{
consensus_table(bamfile, reffile, offset = 1)
}
\arguments{
\item{bamfile}{Path to the BAM file.}

\item{reffile}{Path to the reference sequence (Fasta format).}

\item{offset}{Using \code{\link[GenomicAlignments]{sequenceLayer}}, the refseq and
the query reads seem to be one off. Use this offset to cut off bases from the
beginning of the query reads and the end of the refseq.}
}
\value{
A data.table
}
\description{
Generate a consensus table between a BAM alignment and a reference sequence.
}

