% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{sam2bam}
\alias{sam2bam}
\title{Convert a samfile to a bamfile}
\usage{
sam2bam(samfile, reffile)
}
\arguments{
\item{samfile}{Path to the sam file.}

\item{reffile}{Path to the reference sequence (Fasta format)}
}
\value{
A list containing the paths to a bam file (.bam) and
  an index file (.bai)
}
\description{
Convert a samfile to a bamfile
}

